Firmware update for 1.0.1 including tape and floating point support
This firmware revision brings the source code release up to version 1.0.1 with tape and 32-bit floating point support.

AtMega168 firmware images are now provided as .bin files rather than .hex, because it's not possible to download a new .hex firmware image because they would overwrite the bootloader. Instead, an intermediate firmware image is downloaded; and this one is used to load, from audio, the new proper firmware image which is then programmed into external flash; the AtMega168 USB bootloader is overwritten with the new audio bootloader; which is then run to extract the new 16K firmware image and program it into flash. It's convoluted.

Subsequent upgrades are a little simpler. You download the new audio firmware image, which is a Forth program which downloads the rest of the firmware over audio; programs it into external flash and then runs the bootloader to upgrade, so it's still a one stage process.

These rules don't apply to AtMega328 versions, they can still upload via the USB bootloader, although tape loading is also supported.

What this release doesn't do is update the samples properly, nor is the code tidied, nor does it provide the firmware and other software that supports tape uploading. That will come in a future release.